Prevent user from deleting company when it is used as a reference and display error message

+++ controllers/companies.js
@@ -1,4 +1,5 @@
 import { Company } from '../models/company.js';
+import { Password } from '../models/password.js';
 
 export {
     index,
@@ -24,9 +25,15 @@ async function index(req, res){
 async function show(req, res){
     try {
         const company = await Company.findById(req.params.id)
+        console.log(req.query.error)
+        if(req.query.error === 'invalidDelete') {
+            req.query.error = 'Cannot delete company: has references'
+            console.log(req.query.error)
+        }
         res.render('companies/show', {
             title: `Company: ${company.name}`,
-            company
+            company, 
+            error: req.query.error
         })
     } catch (error) {
         console.error(error);
@@ -68,13 +75,18 @@ async function deleteCompany(req, res){
     try {
         let company = await Company.findById(req.params.id);
         if(company.owner.equals(req.user.profile._id)){
-            await company.delete();
-            res.redirect('/companies');
+            const match = await Password.find({}).where('company').equals(company._id).limit(1);
+            if(match) {
+                throw new Error('Cannot delete company: has references');
+            } else {
+                await company.delete();
+                res.redirect('/companies');
+            }
         } else {
             throw new Error('Not authorized');
         }
     } catch (error) {
         console.error(error);
-        res.redirect('/companies');
+        res.redirect(`/companies/${req.params.id}/?error=invalidDelete`);
     }
 }
